Added Interfaces to Service & Repository
